refactor(api): make the v1 error contract consistent
The v1 reset moved routes and success payloads to the new contract but left
error responses on the old shapes. The API returned four different error
bodies: a camelCase Response<T> envelope from XpenseController, a PascalCase
one from the global middleware, bare 404s with no body, and real problem
details. Tests only asserted status codes, so none of it was caught.

Replace all of it with RFC 7807 problem details:

- Delete the Response<T> / PaginatedResponse / ErrorDetails envelope and the
  XpenseController base that applied it. Controllers extend ControllerBase and
  use plain Ok(), so `new OkObjectResult(...)` opt-outs are no longer needed.
- Group the 23 domain exceptions under NotFoundException,
  PersistenceFailedException and DomainRuleViolationException, and add one
  IExceptionHandler per type plus a fallback. Special cases register ahead of
  their base: insufficient funds stays a field error rather than a flat 400.
- Replace GlobalExceptionHandlerMiddleware, which never awaited its response
  write, with UseExceptionHandler and AddProblemDetails.
- Adopt FluentValidation as the single validation system, run from one action
  filter, and drop the DataAnnotations attributes that were producing a second
  error shape with PascalCase keys.

Fixes found while reviewing:

- GET /api/v1/transactions returned 500. Omitted page/pageSize bound to 0 and
  failed FilterQuery.IsValid(); they now default to page 1, pageSize 25.
- Category create/update failures returned 404, account create returned 400.
  All three are persistence failures and now return 500.
- Timestamps were written with DateTime.Now and read back as Unspecified, so
  they only round-tripped by accident. Everything is UTC now, with a value
  converter tagging reads as UTC. Verified no EF model changes.
- Money's operators discarded the currency: OfCents defaults to EUR, so
  USD + USD returned EUR. Money is now immutable and preserves currency;
  the meaningless Money*Money and Money/Money operators are gone.
- Delete failures threw AccountUpdateFailedException.
- Removed dead code: GetAccountByNumberUseCase, three unused request DTOs.
- Tag update took "name" while create and the response used "label".

Tests: consolidated the six V1*Tests files into one ApiEndpointTests, added
error-body assertions, and seeded a row per resource so the collection
assertions no longer pass against empty arrays. 54 passed, 2 skipped.

+++ src/Xpense/Xpense.Services/Features/Transactions/UseCases/TransferTransactionUseCase.cs
@@ -26,7 +26,7 @@ public sealed class TransferTransactionUseCase(
                 ?? throw new AccountNotFoundException(command.SourceAccountId);
             var destination = await accountRepository.GetById(command.DestinationAccountId)
                 ?? throw new AccountNotFoundException(command.DestinationAccountId);
-            var amount = command.Amount.ToSingle();
+            var amount = command.Amount.ToDecimal();
 
             if (source.Balance < amount)
                 throw new InsufficientFundsForTransferException(source.Id, source.Balance, amount);
